jsonblob: implement Load as documented
