+++ libvuln/updates.go
@@ -26,7 +26,7 @@ func OfflineImport(ctx context.Context, pool *pgxpool.Pool, in io.Reader) error 
 	ctx = zlog.ContextWithValues(ctx, "component", "libvuln/OfflineImporter")
 
 	s := postgres.NewMatcherStore(pool)
-	l, err := jsonblob.Load(ctx, in)
+	l, err := jsonblob.NewLoader(in)
 	if err != nil {
 		return err
 	}
